==> 4_homes_clean.ipynb <==
import numpy as np
import pandas as pd

pd.set_option('display.max_colwidth', None)
pd.set_option('display.max_columns', None)

data = pd.read_csv('home_data_raw.csv')
years = pd.read_csv('years_built.csv')

### Drop duplicates and merge

data.drop_duplicates(subset='detailUrl', inplace=True)
years.drop_duplicates(inplace=True)

df = pd.merge(data, years)
print(df.info())

### Drop and rename colums

df.drop(columns=['Unnamed: 0', 'latitude', 'longitude'], inplace=True)
df.rename(columns={'zipcode': 'zip', 'bathrooms': 'baths', 'bedrooms': 'beds',
                   'livingArea': 'home_sqft', 'homeType': 'type', 'detailUrl': 'url',
                   'years': 'year_built'}, inplace=True)
df.info()

### Change non-date years to np.nan

df.loc[df[df['year_built']=='data'].index, 'year_built'] = np.nan
df['year_built'] = df['year_built'].astype('float')
df.info()

### Drop all 'LOT' home types

df[df['year_built'].isnull()]['type'].value_counts()

df.drop(df[df['type']=='LOT'].index, inplace=True)
df.reset_index(drop=True, inplace=True)
df.info()

### Drop/Replace null and unwanted 0 values

df.describe()

df[(df['baths'].isnull()) | (df['baths']==0) | (df['beds'].isnull()) | (df['beds']==0)]

# Drop samples with no reported 'baths', 'beds', or 'home_sqft'
df.drop([217, 297], inplace=True)

# Find strongest correlations to 'baths' to estimate NaN values
df.corr(numeric_only=True)['baths'].sort_values(ascending=False)

# Use linear regression to estimate 'baths' based on 'home_sqft'
from sklearn.linear_model import LinearRegression

df_baths = df[['home_sqft', 'baths']].dropna()

lr = LinearRegression()
lr.fit(df_baths['home_sqft'].values.reshape(-1, 1), df_baths['baths'])
y_predict = lr.predict(np.array(3554).reshape(1, -1))
df.loc[219, 'baths'] = np.round(y_predict)
df.loc[219]

#### Drop/replace null and unwanted 0 'home_sqft' values

nulls_sqft = df[df['home_sqft'].isnull() | (df['home_sqft']==0)]
nulls_sqft

meds_sqft = df.groupby(['baths', 'beds'])['home_sqft'].median()
meds_sqft

for i in nulls_sqft.index:
    sqft = meds_sqft[nulls_sqft.loc[i, 'baths'], nulls_sqft.loc[i, 'beds']]
    print(sqft)
    nulls_sqft.loc[i, 'home_sqft'] = sqft

nulls_sqft

df[df['home_sqft'].isnull() | (df['home_sqft']==0)] = nulls_sqft
df.info()

### Drop/Replace null 'lotAreaValue'

print(df[(df['lotAreaValue'].isnull()) & (df['type']=='SINGLE_FAMILY')])

# Drop samples with 'buildable plans'
df.drop(df[(df['lotAreaValue'].isnull()) & (df['type']=='SINGLE_FAMILY')].index, inplace=True)

# Set null 'lotAreaValue's equal to 'home_sqft
null_lots = df[df['lotAreaValue'].isnull()]
df.loc[null_lots.index, 'lotAreaValue'] = df['home_sqft']
df.loc[df[df['lotAreaUnit'].isnull()].index, 'lotAreaUnit'] = 'sqft'
df.info()

### Find 'lot_sqft' for all homes

# Make a new column 'lot_sqft'
df['lot_sqft'] = df['lotAreaValue']

# Convert homes with acres to sqft
df.loc[df[df['lotAreaUnit']=='acres'].index, 'lot_sqft'] = df['lotAreaValue'] * 43560
df.describe()

# Set condo and townhouse 'lot_sqft' equal to 'home_sqft'
df.loc[df[(df['type']=='CONDO') | (df['type']=='TOWNHOUSE')].index, 'lot_sqft'] = df['home_sqft']

df.drop(columns=['lotAreaValue', 'lotAreaUnit'], inplace=True)
df.reset_index(drop=True, inplace=True)
df.info()

df.describe()

df.to_csv('home_data_clean.csv')
